USB: serial: ftdi_sio: Add FT230X PID
The BlueFly USB models have a FT230X chip, which is newer than the Kobo
Mini kernel. But, at least for the functionality we require, it is
nevertheless compatible with the 2.6.35.3 ftdi_sio driver.

With this change, reprogramming the BlueFly FTDI EEPROM (which was
BlueFly's recommended solution until now) is not necessary any more.